requirements.txt
ollama
chromadb
langchain
langchain-core
langchain-ollama
langchain-community
langchain_text_splitters
llama-index
langchain-chroma
fastembed
pikepdf
sentence-transformers
elevenlabs
mcp
streamlit
python-dotenv
nest-asyncio
typing-extensions